Proyecto actualizado

+++ Projecto1_DiegoMorales_1132119/Nodo.h
@@ -6,6 +6,6 @@ public:
 	~Nodo();
 	Nodo *Anterior;
 	Nodo* Siguiente;
-	int digito;
+	double digito;
 };
